project: rename cmd/dockerd to cmd/server
