+++ cmd/server/README.md
