+++ cmd/server/config_solaris.go
