+++ cmd/server/daemon_test.go
