+++ cmd/server/daemon_windows.go
