+++ hack/make/.binary
@@ -8,8 +8,8 @@ binary_extension() {
 	fi
 }
 
-GO_PACKAGE='github.com/docker/docker/cmd/dockerd'
-BINARY_SHORT_NAME='dockerd'
+GO_PACKAGE='github.com/docker/docker/cmd/server'
+BINARY_SHORT_NAME="${DAEMON_BINARY_NAME}"
 BINARY_NAME="$BINARY_SHORT_NAME-$VERSION"
 BINARY_EXTENSION="$(binary_extension)"
 BINARY_FULLNAME="$BINARY_NAME$BINARY_EXTENSION"
